Add 'template query' tag to enable queries for custom templates; use it in combination with 'template mode' on uploaded files.
Each query string is interpolated like:  iqs = qs % self.storage
So it can incorporate named query parameters from the request URI
into the evaluated query string.

Therefore, the quoting rules are clumsy:
  1. use %(key)s for each query parameter 'key'
  2. use %% for each literal % in the base URI, e.g. in quoted fields
  3. use percent-encoding in transport layer, e.g. REST API

The results of each query are evaluated and stored into a dict:
   self.globals.template_query_results[qs] = results

So a custom template 'name=demo' could be tagged with a query like:

   name=%(name)s;tag1=quoted%%20value(tag2;tag3)

and a request can be:

   GET /file/name=demo?name=foo%2520bar

which will cause the following interpolated query string:

   iqs = qs % { 'name' : 'foo%20bar' }

e.g.

   name=foo%20bar(tag1=quoted%20value(tag2;tag3)

which will be evaluated like a sub-query (using the general query-path
evaluator). The custom template can find the query result list under
the following global:

   template_query_results['name=%(name)s;tag1=quoted%%20value(tag2;tag3)']

i.e. it can use the static query string (before interpolation) as the
dictionary key.

Note: the projection feature of the query string, e.g. (tag2;tag3) can
be used in the query string to fetch extra tag metadata for the
matching results, as usual.