merge Python test script with conn reuse r4965 from branches/denorm
benchmarking script reimplemented in Python for HTTPS connection reuse

usage:  

   export LOAD_FILE="test-data.csv"
   export LOAD_TAGS="tag1;tag2;tag3;..."

   export TEST_STRIDE=10000
   export TEST_SIZES="1 10 100"

   python tagfiler-benchmark.py 

The LOAD_TAGS are a properly URL-escaped list of tagnames such as
could be embedded in a /subject/tag1(tag1;tag2;...) URL, and the
script will pull off the first tagname and use it as the subject
identifying tag.

The LOAD_FILE is a CSV formatted file with a column for each of the
respective LOAD_TAGS and one record per line.

The TEST_STRIDE configures the number of records inserted between each
major test round, e.g. to grow the graph for a scalability test.

The TEST_SIZES configures one or more subsets of the stride data to be
probed at each major test round, e.g. to time the latency of a
PUT/GET/DELETE of that many rows for a given total graph size.

The output to stdout is a CSV table with header such as:

hostname,graphsize,PUT 1,GET 1, DEL 1,PUT 10,GET 10, DEL 10,PUT 100,GET 100, DEL 100
crater.isi.edu,0,0.387,0.033,0.106,0.383,0.033,0.116,0.421,0.037,0.183
crater.isi.edu,100000,0.392,0.032,0.177,0.416,0.032,0.191,0.427,0.041,0.261

where times are fractional seconds.