modularize the predlist path query generator a bit, add new rangemodes
Rangemode 'tagvalues<' and 'tagvalues>' return a table of format:

  tagname, value, count

for each of the list predicates, sorted by ASC or DESC count for
tagvalues< and tagvalues>, respectively.  Because the result table
must be monomorphic even when combining different tags with different
dbtypes, all values are converted to text for output.

The value set per tagname is also limited by the limit query param, so
you can get the top or bottom N values per tag.

This differs from the 'values<' and 'values>' rangemode where only a
single row is returned and the count information is stripped out.