implement defaults=column,... query parameter for /entity/ POST API
one or more column names can be given i.e. 

   POST /ermrest/catalog/X/entity/tabename?defaults=column1
   POST /ermrest/catalog/X/entity/tabename?defaults=column1,column2

to tell server that the client wishes to get server-side defaults
applied to those columns, for example to get auto-numbered serial
columns to work.

If columns are all non-metakey columns, they simply get a default
value while all other processing is the same as usual.

If any columns are metakey columns, then all key testing is deferred
to the database, assuming that each input row will produce a new
non-conflicting row.

This is only allowed on POST since the "upsert" logic of PUT is
impossible to express if all metakey material is not known for record
correlation.