enable generalized filter conjunctions and fix related URL parsing bugs on filters
1. the outer path syntax for filters remains .../filter and a sequence
   of such /filter suffixes are a conjunction

2. the core syntax of a filter is the predicate:

     column_name UNARY_OP

     column_name BINARY_OP value

   where the only current UNARY_OP is '::null::'

   and the current BINARY_OPs are: '=', '::geq::', '::gt::',
   '::leq::', '::lt::', '::regexp::', '::ciregexp::', '::ts::'

3. a single filter can combine multiple predicates and negation using
   these metacharacters in order of decreasing precedence.

   A. negation as: '!' predicate

   B. conjunction as: predicate '&' predicate

   C. disjunction as: predicate ';' predicate
   
4. filters can use parenthetical grouping to override precedence of
   negation, conjunction, and disjunction metacharacters:

   '(' predicate ')'

   but any such expression must appear completely within a single
   filter path element.

Using these rules, and assuming P1..PN are basic predicates with
optional '!' prefix, we can illustrate different filter construction
idioms.

A logical function can be written in disjunctive normal form as a
single filter element in the path using top-level disjunction to
combine a sequence of conjoined predicates:

   /P1&...;P2&...;...;...&PN

A logical filter can be written in conjunctive normal for as a series
of conjoined filter path elements each containing disjoined
predicates:

   /P1;.../P2;.../...;PN

And finally, using parentheses a logical function can be expressed as
a single nested filter expression in denormalized form, including with
top-level negation if desired:

   /(...)

   /!(...)

this may better match an application developer's conceptual
understanding of the logical function.