This check-in includes some refinement on the use of AngularJS filters.
There are (at least) two ways to bind data in AngularJS: the "double-brace"
syntax and ng-bind directive. It wasn't clear that ng-bind could use filters in
the same way. The previous revision injected the filter into the controller for
use with ng-bind, but this is not necessary. This revision uses the same
piping-style for each method of data binding.