enable nested-object retrieval in attribute, attributegroup, aggregates
in the attribute slot of the APIs, allow an alias-qualified '*' to
imply a nested record containing the entire row of the table
designated by that path alias.

for example:

  GET /attribute/A:table1/table2/id,subrow:=A:*
  Accept: application/json

this will return something like:

[{"id":1,"subrow":{"field1":1,"field2":2}}...]

assuming tablel2 has a column "id" and table1 has columns "field1" and
"field2".

this also works with the array() aggregate function to get back an
array of related rows:

  GET /attributegroup/A:table1/table2/id,subrows:=array(A:*)
  Accept: application/json

this will return something like:

[{"id":1,"subrows":[{"field1":1,"field2":2}...]}...]

output of this new feature for text/csv content type is idiosyncratic
and subject to changes and improvements.  With JSON encoded output,
the subrecord or array should appear as nested parts of the outer JSON
record as illustrated above.