Re-implemented how the "Print List" button works. Instead of sending the results table to a popup window, loads the results in a new print screen before sending to host print dialog. This removed the "Details" buttons from the printouts, but lets the "Logout" button show at the top of the page. (Where it can be easily ignored when reading the printout.)
After the print job is submitted (or canceled), the results page is
restored. Various functions were added to the services.js and controller.js
files to support this change.

Button placements are more regular. Button with the same or similar
functionality show in the same general location on the screen.

Initialization of the AngularJS controllers is now only in the ngRoute
configuration. Earlier revisions had mistakely included this initialization in
both the ngRoute configuration and HTML templates. This double initialization
is usually transparent to the user, but needed to be fixed anyway.