add proof-of-concept volume slicer and ray-caster tools
these require the PyOpenGL and vispy libraries which you can try to
install from pip:

   pip install PyOpenGL
   pip install vispy

they both take a single command-line argument which should be a TIFF
dialect such as LSM or OME-TIFF.

they may fail and report many errors if the image requires a larger
3D texture size than is supported on your platform.

when they work, they both draw the volume data with an automatic slow
rotation on the screen.  the window can be resized but there is no
other UI support.

hit the escape key to exit.