add "/$alias" entity-path context syntax
This syntax allows you to set the entity-path "context" to an earlier
element in the path, rather than the default which is to always be the
final element.  For example:

 GET /entity/Foo
 GET /entity/A:=Foo/$A

these are equivalent and both address Foo entities.  Where it gets
interesting is in building context and then resetting:

 GET /entity/Foo/a=2/Bar/x=5
 GET /entity/A:=Foo/a=2/Bar/x=5/$A

The first returns Bar entities where Foo.a=2 and Bar.x = 5.  The
second returns Foo entities under the same conditions.  You can also
reset context in the middle of a path to affect resolution of
subsequent filters and entity links:

 GET /entity/A:=Foo/B:=Bar/$A/a=2/$B/x=5
 GET /entity/A:=Foo/B:=Bar/$A/a=2/$B/x=5/$A

This pair is equivalent to the previous pair, respectively.  You can
therefore build context trees that were not previously possible:

 GET /entity/A:=Foo/a=2/Bar/x=5/$A/Baz/z=N/$A/Bam/j=1/$A

This returns Foo entitities linked to Bar, Baz, and Bam entities with
their respective filters, i.e. something like this SQL:

  SELECT "A".*
  FROM "Foo" "A"
  JOIN "Bar" "B" ON ("A"... = "B"...)
  JOIN "Baz" "C" ON ("A"... = "C"...)
  JOIN "Bam" "D" ON ("A"... = "D"...)
  WHERE "A".a = 2
    AND "B".x = 5
    AND "C".z = 'N'
    AND "D".j = 1