Registry, CatalogFactory Catalog, and sanepg2 now use libpq DSN connection strings.
The database configuration of ERMrest and its catalogs are now based on libpq 
DSN strings (e.g., 'host=... post=... user=... password=... dbname=...'). In
the simplest form 'dbname=' it is equivalent to passing the empty string to 
the dbname parameter. It will connect over domain socket to the database named
after the database user.

This change requires an update of the ermrest_config.json. It should be as 
simple as chaninging dbname='' to dsn='dbname='. See the new template for an
example.

This change should be backward compatible with existing catalogs registered in
the SimpleRegistry implementation.

To configure ERMrest to use a PostgreSQL server over TCP/IP, just extend the 
DSN strings with the appropriate parameters. Also, extend the "catalog_factory"
 "template" parameter with similar key-value pairs. They will be used in the
catalog's "descriptor" and serialized into a libpq DSN string.

Clearly, PostgreSQL pg_hba.conf also needs to be updated to allow password 
authentication.