Fix regression on model changes introduced recently by /textfacet/ API hacks
The matrialized view _ermrest.valuemap creates dependencies on nearly
every model element, so it needs to be dropped before we can delete
tables, schemas, columns, etc.

With this change, the test/rest-tests.sh suite can succeed again, when
called with an existing catalog in the TEST_CID environmental
parameter.  Further work is still needed to allow multi-tenant catalog
creation to work because the /textfacet/ API broke that too!