prevent valuemap view from being regenerated on every postgres introspection
adds util.view_exists() recognizing regular and materialized views, and
uses this for conditional rebuild of valuemap.  previous code was looking
for a table which always would return false!

also, emit a debug log for now to make it easier to see how often we are
rebuilding the valuemap; this code needs to be improved soon.