refactor connection pooling slightly to add PooledConnection w/ final() method
The previous, partial refactoring abused the term 'dsn' to mean either the
database dsn string or a triple of (pool, conn, cur) which was a big mess.

Now, use PooledConnection(dsn) instead of pooled_connection(dsn) to
construct an object instead of triple.  Use its perform(...) method
instead of pooled_perform(dsn, ...), and add a final() method that should
be called whether or not perform() gets run.

The use of final() fixes a database connection leak that occurred for
some under tested cases where NoMethod() or other exceptions were
raised outside of the perform() code-path.