Properly match key URIs in annotation and add regression test [fix #37]
Setting an annotation via key worked for the first key URI on a given
subject, e.g. a table, but then setting a second annotation was
treated as an update of that one even if a different key URI was
specified.

Also extended rest-tests to use multiple keys and cross-check results
of PUT, GET, and DELETE more completely for multiple keys and values.