add test case to exercise composite key issue from #38
A table as below with composite key seems to cause ERMrest to fail internally with an odd error message.

CREATE TABLE test (
id int8,
last_update timestamptz,
name text,
site int8,
PRIMARY KEY(id, site)
);

A PUT to the /entity/test API fails with a message like:

The request conflicts with the state of the server. UNION types bigint and timestamp with time zone cannot be matched.

The UNION is internal to the ERMrest server logic and seems to have a column ordering expectation that is not met between the update and insert CTEs in the emulated upsert query.