rework paging modifiers to correct bug in before-key handling
- Only one of `@before` and `@after` is allowed in one request
- `limit` parameter is required when using `@before`
- Updated docs for feature

The desired semantics are to get the rows *immediately* before or
after the page key.  To get after, we can just use the page key as a
lower bound on a query and send arbitrary numbers of results. To get
before, we have to run a query with reversed sort order, stop when we
get up to the *limit* number of rows, then re-sort with the client's
desired sort order.