fix double-array bug in entity PUT with JSON response content type
A recent code change appends two query results into the output stream
but this needed to be tweaked because the queries are already
serialized and each has its own `[...]` framing when serving a JSON
response.  When one of the intermediate results is an empty array
we can discard it and return the other result. When both are non-empty
we need to trim off the framing and add a comma separator instead.