set comments on keys and fkeys when provided in JSON input, add test cases
A comment field in the JSON input should apply the comment when
creating new keys and foreign keys either on existing tables or as
part of a newly posted table with embedded sub-resources.

There was a missing fkr.set_comment() call in the fkr.add() method.

The key management had to be refactored to look more like fkr
management so that it could preassign a constraint name and then set
comment on the named constraint, rather than having postgres assign an
unpredictable constraint name.