bypass model cache for model change API to fix inconsistency hazard
- The model is a complicated graph of objects
- When processing model change requests, we update model as we go to help with producing success output representations
- On commit, we update DB schema and _ermrest.model_version tracking
- On rollback, this would leave the model graph inconsistent with DB

Most requests are read-only as far as model is concerned and benefit
from a cached (in Python) model graph that is transactionally
consistent. But for the model-change API methods, we can just run
introspection again to produce our own private copy, mutate it at
will, and discard it at the end of the operation regardless of
outcome.

After failures, any cached model is unchanged. After sucess, the model
version tracking will invalidate the cache and we will introspect
again anyway.