Add outer-join support, docs, and basic tests
This adds optional prefix to the explicit column-linkage notation:

- `/attribute/A:=S:T/(id)=(S2:T2:fk)/$A/A:*,*`: Backwards-compatible inner join
- `/attribute/A:=S:T/left(id)=(S2:T2:fk)/$A/A:*,*`: Left outer join
- `/attribute/A:=S:T/right(id)=(S2:T2:fk)/$A/A:*,*`: Right outer join
- `/attribute/A:=S:T/full(id)=(S2:T2:fk)/$A/A:*,*`: Full outer join

Note that all these joins are subject to the same reduction rules as
before, so they are not identical to a basic SQL query.  You have
to remember:

- `/entity/` and `/attribute/` reduce join combinations to distinct path context records
- `/aggregate/` reduces join combinations by aggregate functions
- `/attributegroup/` reduces join combinations to distinct group keys

So each example query above could have different row counts if you omit the `/$A` context-reset element.