fix mangling of values with `%` character in them
The underlying sql_identifier and sql_literal utility functions were
applying a bogus transform that replaced `%` with `%%`. This was
previously necessary with the web.database backend but should have
been removed when we switched to psycopg2.

In the unlikely event that someone successfully used a name with `%`
in it and got the old codebase to create it in schema with `%%`, they
will now encounter a reverse inconsistency. The DBA would have to
intervene and correct the actual catalog schema to use a single `%`
where it was originally intended, replacing the corrupt `%%`.

However, under practical scenarios, I think the old codebase would
have failed to allow one to create and use a schema, table, or column
name with a `%` in it due to inconsistent behavior of the REST API for
schema management and later access.

With this patch, end-to-end use of such names is now possible, and
should also be consistent if a DBA has locally created names with `%`.