don't try to skip idempotent data updates to tables
This was meant to be faster for redundant updates, but causes correctness
issues.

1. The response messages are confusing to clients as rows are missing
2. The skipped rows cannot be handled by triggers that might have otherwise changed the updated row content!

I think this trigger scenario is more too important, so the
performance regression is necessary. Deployments depending on fast
idempotent insertion where they keep sending the same data are just
going to have to live with the costs or be more careful about
redundant data submissions for large tables.