fix internal server error when there are no columns populated by entity POST
This can happen if the client lists all columns in the defaults list.

Less likely, it can also happen if a client tries to insert data into
a zero column table!

Turn these into 409 Conflict errors since it is a conflict between the
request and the current table definition.  If the table had an additional
column defined, it might work to send the same request...