add bulk schema updates on catalog/N/schema resource
Two forms:
1. `{"schemas": { schema name: schema representation, ... }}`
2. `[ schema or table representation, ... ]`

The first mirrors the result of GET on the same resource and can
create any number of new schema and their constituent tables all at
once.

The second can mix individual schema (with or without embedded tables)
and table representations.  The real purpose of this form is to create
new tables in an existing schema.

To resolve any ambiguity, the second form requires processing of the
list of representations in document order, so if a table is lifted out
of its schema rather than embedded, the schema must already exist in
the catalog or must appear as an earlier element in the list before
the table which is meant to be added to that schema.

Both forms support circular dependencies between newly created tables
by applying the model elements in two phases: first, schema and table
elements are added; then, foreign key constraints are added to all the
new tables so that they can reference any other table from the same
batch independent of insertion order.