propose 2016:foreign-key annotation field `domain_filter_pattern` [ci skip]
This pattern can be expanded via mustache to produce an ERMrest
filter string.  It can interpolate existing column values to apply
dynamic filters, and can also embed static filters.

The filter string can be appended to an entity-path to filter the
elements of the domain table referenced by the foreign key. This
allows a DBA to present a subset of choices for data-entry on columns
governed by a foreign key.

For simplicity, specify a limited subset of ERMrest filter syntax:
- no `/` so no joins
- no alias qualifiers so only flat filters on the domain table
- conjunction via `&` only
- disjunction via `;`
- parentheses and negation
- the usual operators

This language can be validated for basic safety using a regular
expression but requires a context-free grammer to fully analyze.
- A naive client can just do a safety check with regexp match
- The filter can be inserted into a domain query URL template
- If ERMrest rejects it with bad request or conflict errors, consider it bad
- Since no `/` is allowed, it cannot change the fundamental structure of the basic domain query

The safety validation regexp would look something like:

    ([=!();&]|::[a-z]+::|([A-Za-z0-9-_.~%]|%[0-9A-Fa-f][0-9A-Fa-f])+)*

in other words, it only allows permutations of the basic single-char
combining punctuation, special `::` delimited operators, and encoded
strings with non-reserved characters or `%hh` hex-encoded triplets.