change textfacet api from precomputed valuemap to dynamic union query
This will be necessary to allow per-row or per-column authz decisions
to affect value visibility in the upcoming ACL features.

1. Use a long chain of UNION ALL queries producing individual schema, table, column triples.
2. Can reuse the existing per-column trigram indices established for regular regexp matching.
3. Initial (incomplete) prototype of local policy to decide which schemas, tables, columns to include in textfacet searches.

This patch can cause severe performance regressions when mixed with
postgres row-level security policies. These policies can produce very
inefficient query plans as compared to an equivalent query where the
decision logic is embedded in the conjunctive WHERE clause.  So, we
keep this in the new feature branch until ACL support is sufficient
to replace row-level security in practice.