add table-level mutation ACL test permutations, bug fixes, refactoring
Test each of these scenarios:
- insert only (no update, delete, or select)
- insert and select allowed (no update or delete)
- update and select allowed (no insert or delete)
- all writes allowed

To make these behave as desired, changed a few other bits:
- defer enforcement of select until AnyPath.get() to allow insert-only
- specialize Table.has_right() and Column.has_right() to check parent rights
- drop table and schema-level checks from various has_right() in model.key

The current behavior is still a bit out of sync w/ docs on column
access rights. The docs propose a more baroque model with sub-types of
column mutation, while the current code just has 'update' which aligns
with table-level 'update'.  We may want to just narrow the scope in
the docs since we don't have strong enough use-case motives for the
other access rights.