add static rights sub-documents in model introspection and test cases
This is not an extra REST API but just a field in the existing model
resources, called `rights` and keyed by ACL names with a boolean
decision for each.  Two ACLs are omitted:

- `enumerate`: all visible model elements have `True` enumeration decision
- `write`: this is just a short-hand for the actual rights `insert`, `update`, and `delete` which are advertised individually.

The test cases only go through catalog, schema, table, and column
levels and only for the policy configurations currently tested in
the authz_static test module, which could benefit from more coverage
in general.