update acls.md replacing projection strings with projection documents
Define an abstract syntax tree using JSON for a supported subset of
query features, rather than having open-ended query fragment language
and needing a fragment parser.

- JSON list for a path
- ends with bare column name string to project
- link element to join by foreign key only (no ad hoc joins)
    - name fkey by constraint name pair `[schemaname, constraintname]`
    - specify direction of fkey linkage *inbound* or *outbound*
    - optional *context* to link to aliased table instance
    - optional *alias* to bind alias to new table instance
- filter element to constrain column
    - alias-qualified or bare column references
    - operator name defaults to `"="`
    - operand for binary operators
    - optional negation of result
- and/or elements to combine filters
    - list of filters combined by combiner
    - optional negation of result

With the foreign key names exposed, we could store those in the
database and use them to detect conflicts such as trying to remove an
fkey that is needed for the active policy.