partial fixes for dynacl enforcement subqueries [ci skip]
- pass `prefix` into subquery joins to get correct table aliases
- introduce `enforce_client` to selectively disable authz checks
    - defaults to `True` for main data query
    - disabled for authz sub-queries so "hidden" policy can be consumed
    - might need other safety checks here...

This code revision is overly trusting of policy authors.  A
sufficiently complex dynamic ACL could create a covert channel to
discover properties of tables the policy author and client do not
otherwise have rights to see.  It's not a risk for catalogs where only
trusted DBAs have ownership over the catalog, schemas, and tables.

If we did not disable enforcement of client rights for the policy
sub-queries, we would get a situation where these indirectly accessed
policy tables can cause Forbidden/Unauthorized errors even if they are
not needed for a particular access solution.  It seems likely for many
projects that some policies may depend on information that would not
be exposed to the less privileged client subject to the policy.

A better balance of trust might be to enforce that the base table
owner (the policy author) has rights to access the content they
reference in their policy.  However, this would require storing the
owner's attribute list at time of policy creation, so we can
cross-check it at runtime when a different client is accessing the
base table and requires policies to be compiled.  This would redefine
the `enforce_client=False` case to instead enforce against the policy
author.

If a policy uses tables that are not allowed for the policy author, it
could be trimmed from the effective policy to allow other more basic
policies to potentially satisfy the client access.