add basic histogram support via new bin(col;lower;upper;nbins) projection
The client can get data for a histogram of numerical column values
by using this as follows:

- .../attributegroup/path/bin:=bin(column;lower;upper;N);cnt:=cnt(*)

The original value range from _lower_ (inclusive) to _upper_
(exclusive) is divided into N equal-width bins. This currently only
works for numerical column types, e.g. integer, float, numeric, and
serial.

The result of such a query will have at most N+3 rows, each of the form:

    { "bin": [ X, L, U ], "cnt": K }

where _X_ is a bin number, _L_ is a bin lower bound, _U_ is a bin
upper bound, and _K_ is the number of rows in the bin.

- _X_ = `null`: The bin for all NULL valued rows
- _X_ = 0: The bin for all rows less than the _lower_ bound
- _X_ = 1: The lowest bin in the range
- _X_ = _N_: The highest bin in the range
- _X_ = _N_ + 1: The bin for all rows greater than the _upper_ bound

If the client does not care about counting NULL or out-of-range
values, they MAY exclude them using the normal ERMrest filter syntax.

Future enhancements might be able to add support for more types or
make the _lower_ and _upper_ arguments optional...  but this version
is usable now for a stateful client application.