add primary key enforcement in introspection and rest api
This adds several checks to ensure that all ermrest tables/views have
valid primary keys, i.e. a unique constraint involving not-null
columns.

1. Fail on introspection w/ 500 Internal Server Error
2. Reject new table definitions w/ 400 Bad Data
3. Reject deletion of final not-null key with 409 Conflict

This enforcment can be disabled with an ermrest_config.json setting:

    "require_primary_keys": false

This new setting is set to `true` in the new sample config and also
defaults to `true` if not configured, e.g. in legacy deployments that
update to this latest code.

For locally defined SQL views, you must use the recently added psuedo
not-null constraint along with existing pseudo key constraints to
satisfy this new strictness test.