changes to 2017-asset annotation
First, I felt the discussion of 'default heuristics' belonged under the category of 'chaise implementation notes' so I moved those.

Based on a discussion with Mike, I added `source_filepath_column` and `path_regex`. The latter might need a better name. These would be useful for some functionality that the deriva-uploader (aka iobox) client could perform following introspection.

I updated the implementation notes adding a new section for the deriva-uploader and appending to the chaise section, to identify which attributes of the annotation are supported by the respective clients.