include system columns in result of POST /entity/ API
These were excluded from the internally-generated `mkcols` and
`nmkcols` and added to the `?defaults` param list implicitly, to make
transitioning to the new system columns easier. But, as a side-effect
this also dropped the system columns from the result of a `POST`
request, meaning the newly issued `RID` value was not conveyed to
the client.