fix input of empty arrays for array column types
The previous code was incorrectly collapsing empty arrays into NULL
due to the way it has to unpack a JSON representation, cast the
elements to the column element type, and repack as a native array.

This fix adds an explicit coalesce() statement for the non-null
input case and adds test cases.