simplify history amendment api proposal [ci skip]
Just use RID of subject model resource with a single polymorphic API
instead of separate APIs for each type of subject resource.

This depends on numeric string versus text tokenization to distinguish
them in the URL parser:

- /ermrest/catalog/N/history/from,until/acl
- /ermrest/catalog/N/history/from,until/annotation
- /ermrest/catalog/N/history/from,until/RID/acl
- /ermrest/catalog/N/history/from,until/RID/acl_binding
- /ermrest/catalog/N/history/from,until/RID/annotation

The first two have the catalog as subject, while the others are for
sub-resources such as schemas, tables, columns, and constraints.