improve catalog-level acl api behavior to match docs
Conceptually, we've said a catalog ACL cannot be `null` since it is
the root of inheritence for sub-resources. Rather than reject a
few requests that could produce a `null` state in the DB, we can just
remap them to apply `[]` instead.

- PUT of a `null` into `acl` or `acl/name` resources
- DELETE of `acl/name` resources

This means that a subsequent fetch of the ACL state will always
include the full list of ACL names, unlike on sub-resources where the
set of ACLs is sparse.

Also changes test cases slightly to match this new expectation.