wip snapshot: add data redaction mechanisms
- `GET /ermrest/catalog/N/history/,/attribute/RID` tells you when the column with _RID_ exists in history
- `DELETE /ermrest/catalog/N/history/from,until/attribute/RID` redacts column in all tuples *contained* in history range
- `DELETE /ermrest/catalog/N/history/from,until/attribute/RID/RID2=J` redacts column in all tuples *contained* in history range if the filtering column _RID2_ has the value _J_ which is a JSON value in the URL
   - _RID_ and _RID2_ must be columns of the same table

This covers scenarios such as:
1. Redact a column entirely for all rows in history range
2. Redact a column only for specific row specified by `RID`
3. Redact a column whenever it holds a certain value

We may want to improve it to also handle historical tuples which *overlap* the specified history range but are not fully *contained*. The current code does not redact them. An enhancement might split the historical tuple into segments before, during, and/or after the redaction history range.