wip snapshot: ugly script to convert a catalog from int8 to base32 RIDs
This is not intended to cover all possible cases but to help
convert a practical subset.  This script may be enabled temporarily
as part of the ermrest-deploy sequence, and then disabled again
before the pull-request is finalized.

This script is only relevant for experimental users who have been
operating catalogs with the development code in this branch and wish
to carry forward accumulated catalog content. It attempts to convert
the live schema and all the history storage and historical schema to
the text-based RIDs while preserving other data, as if it had always
been using base32 text RIDs.

This branch was pre-announced as not providing backwards compatibility
until the pull-request is finalized.  So, this script is purely a
convenience for developers and testers and not a supported production
technique.