for history amendment, synchronize live tuple metadata
We don't want amendment to affect the live data in general, but we
need to adjust the "RMT" field to match whatever changes we've
made to history tuples, so there are no inconsistencies where a
live tuple appears to partially overlap a history tuple.

We also update the "RMB" to show the admin who made the amendment,
to be consistent about who made the changes at "RMT".