refactor acl documentation [ci skip]
Must of the information in user-doc/acls.md actually describes the
current interface and is moved to api-doc.

The api-doc is also slightly simplified to give more compact treatment
to annotations, comments, acls, and acl_bindings as generic
sub-resource models associated with multiple subject resource types.

A few future enhancement sketches and other bits remain in the old
user-doc file.