fix inappropriate 409 conflict error for POST to table with only RID key
The RID key has special handling and gets omitted from the "metakey"
used for upsert logic in the `/entity/` API.

If the table has no other keys, the server was refusing to process a
`POST` request against `/entity/T?defaults=RID` with an error message
that is applicable to PUT requests but not POST.

This change slightly rearranges the guarding tests and adds a test
case allowing the POST request to proceed.  It does **not** however
remove the existing implementation limit where POST requests must
provide at least 1 non-defaulting column.