add `heal_existing` boolean arg to several stored procs and change default behavior
- `_ermrest.model_change_event(heal_existing)`
- `_ermrest.rescan_introspect_by_oid(heal_existing)`
- `_ermrest.rescan_introspect_by_name(heal_existing)`
- `_ermrest.enable_table_histories(heal_existing)`
- `_ermrest.enable_table_history(table_rid, heal_existing)`

These all retain a backwards-compatible variant that lacks the
`heal_existing` parameter and behaves as if it was set `false`.

History tables will be created on demand with the trigger needed to
maintain content. But, if the history table and trigger already
existed, it will be assumed to be clean and up to date unless
`heal_existing` is set to `true`. When set to `true`, exhaustive
history-healing will be performed to seal off any open history tuples
which are no longer present in the live table as well as to replicate
any live tuples not yet present in the history table.

The old default behavior implicitly acted like `heal_existing` which
is too expensive for typically clean redeployment scenarios.

An administrator who has performed SQL manipulation behind the back of
ERMrest, particularly with triggers disabled or combining model and
data changes, should consider using the expensive healing process ot
return the catalog to a self-consistent state.