add support for vocabulary id/uri columns and defaulting behavior
- Two new text domains
   - ermrest_curie meant to indicate storage of CURIE values
   - ermrest_uri meant to indicate storage of URI values
- Magic templating for domains w/ magic column names
   - for any table
   - ermrest_curie on `id` column only
   - ermrest_uri on `uri` column only
   - regexp match for `{RID}` causes replacement with row's RID
   - but only during row INSERT
- Also fixes inconsistency in use of _ermrest.maintain_row trigger
   - Was enabled on tables during deployment already
   - Was missing from tables created by REST API

The new templating feature extends the maintain_row trigger with a bit
of dynamic introspection. If this is found to cause too much overhead
on row insertion, we may need to revisit this implementation choice
and do something uglier to configure a specialized trigger only for
the vocabulary tables that need it.