change introspection-time ACL binding validation to prune invalid bindings
The previous code would expose a 409 Conflict which made the service
inoperable once the user managed to break the model state. This can be
done via REST API, e.g. via removing foreign keys or tables used
in an ACL binding projection path.

This commit modifies the introspection to catch the conflict on a
per-binding basis and perform cleanup. It logs a trace message and
deletes the ACL binding from the stateful model.

Because this cleanup is happening at a place that would normally be
read-only, it commits the transaction and raises a 503 Service
Unavailable error to abort the normal request processing.  A properly
configured client will retry their request and make progress because
the inconsistent part of the model has been pruned. All ACL bindings
are validated or pruned in a single pass, resulting in one 503
response and retry cycle.