stop using expensive exclusion constraints on history tables
These were meant to document and enforce an integrity constraint for
the disjoint intervals for each row's history tuples. However, the
overhead is too high for the actual history-collection trigger.

The previous exclusion constraint on `RID, during` is replaced with a
simpler uniqueness constraint on `RID, during` which could still
catch blatant bugs but leaves a simple b-tree index which is lower
overhead for normal, successful function.

This patch also includes an upgrade mechanism to detect the old
exclusion constraints and replace them on existing history tables,
during the normal redeploy process.