adjust _ermrest.maintain_row() trigger function to guard against missing columns
When require_primary_keys is false in the server config, we allow tables
to be created without all the system columns.  But, the maintain_row()
trigger function was blindly writing to all the system columns.

This patch adds a pretty cheap guard for arbitrary combinations of
system columns by converting the NEW or OLD record to jsonb and using
jsonb operators to test for presence of columns as jsonb object keys.